Query nuget for the download uri
Rather than hardcoding the uri to download PostgreSql.Binaries.Lite from
we use the nuget client SDK to load the users settings and query nuget
for the uri. This will allow the downloader to work even when nuget is
set up to use proxies, mirrors, local caches or other scenarios
(anything that nuget.exe install supports this should now support).